refactor nav
